requirement.txt
appnope==0.1.0
astroid==2.4.0
attrs==19.3.0
backcall==0.2.0
bleach==3.1.5
boost-py==0.0.2
bpy-build==1.1.8
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cmake==3.17.2
cmake-generators==1.0.8
colorama==0.4.3
cryptography==2.9.2
Cython==0.29.20
decorator==4.4.2
defusedxml==0.6.0
docutils==0.16
entrypoints==0.3
filelock==3.0.12
flake8==3.8.3
flann==1.6.13
future==0.18.2
future-fstrings==1.2.0
gitdb==4.0.4
GitPython==3.1.1
idna==2.9
importlib-metadata==1.6.0
ipykernel==5.3.0
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.17.1
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.5
jupyter-console==6.1.0
jupyter-core==4.6.3
keyring==21.2.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.4.0
nbconvert==5.6.1
nbformat==5.0.7
nose==1.3.7
notebook==6.0.3
numpy==1.19.0
packaging==20.3
pandocfilters==1.4.2
parso==0.7.0
peppercorn==0.6
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pkginfo==1.5.0.1
pluggy==0.13.1
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
ptyprocess==0.6.0
py==1.9.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.6.1
pylint==2.5.0
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==5.4.3
python-dateutil==2.8.1
python-pcl==0.3.0a1
pywin32-ctypes==0.2.0
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==4.7.5
QtPy==1.9.0
readme-renderer==26.0
requests==2.23.0
requests-toolbelt==0.9.1
rope==0.16.0
Send2Trash==1.5.0
six==1.14.0
smmap==3.0.2
svn==1.0.1
tensorboardX==2.0
terminado==0.8.3
testpath==0.4.4
toml==0.10.0
torch==1.5.1
torchvision==0.6.1
tornado==6.0.4
tqdm==4.45.0
traitlets==4.3.3
twine==3.1.1
typed-ast==1.4.1
urllib3==1.25.9
vtk==9.0.1
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.1.0
